requirements.txt
flask==3.0.0
gunicorn==21.2.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
werkzeug==3.0.1
reportlab==4.0.7
pillow==10.1.0
